Added initial support for sending events to Amazon EventBridge.
The handler takes following optional flags and environment variables:

--endpointId/-e
--clusterArn/-c
--eventBusName/-b

KW_EVENTBRIDGE_ENDPOINT_ID
KW_EVENTBRIDGE_CLUSTER_ARN
KW_EVENTBRIDGE_EVENT_BUS_NAME

endpointId - The URL subdomain of the endpoint. For example, if the URL
for Endpoint is https://abcde.veo.endpoints.event.amazonaws.com, then
the EndpointId is abcde.veo.

clusterArn - The ARN of the EKS cluster where kubewatch is deployed.
This is used to populate the Resources field of the PutEventsRequestEntry
object when calling the PutEvents API for additional context.

eventBusName - The name or ARN of the event bus to receive the event.
Only the rules that are associated with this event bus are used to match
the event. If you omit this, the default event bus is used.